requirements.txt
transformers
accelerate
diffusers
gradio
devicetorch
chatterbox-tts
scipy
librosa
accelerate
peft
